Extract reply_retry_types.rs from turn.rs (parent #680)
Move the assistant-reply retry state types out of `turn.rs` into a
focused sibling module so `turn.rs` keeps shrinking toward
responsibility-focused units (parent meta-issue #680).

One struct + one enum + one constructor extracted (no impl Agent):

- `AssistantReplyRetryState` (per-attempt accumulator: 7 bool/usize
  fields including `downgraded_native_tools`, `retries_remaining`,
  `tool_call_format_retries_remaining`,
  `extra_transport_retries`, `transport_retry_count`,
  `focused_edit_timeout_retry_count`,
  `tool_call_format_retry_count`)
- `AssistantReplyRetryDecision` (`Retry` / `ReturnReply(reply)` /
  `Fail(err)` triplet)
- `AssistantReplyRetryState::new(chat_retries, message_count)`
  constructor (doubles the transport budget on long conversations:
  4 vs. 2)

Behavior unchanged: byte-for-byte identical field defaults and
`message_count >= 12` branch in the constructor.

`pub(super)` limited / no facade re-export (DR3-001). Only consumer
is `reply_retry.rs` — import path updated from
`super::turn::{AssistantReplyRetryDecision, AssistantReplyRetryState, USER_INTERRUPT_ERROR}`
to `super::reply_retry_types::{...}` + `super::turn::USER_INTERRUPT_ERROR`.

LOC delta:
- turn.rs: 99 → 68 (-31)
- new module: +51

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
