+++ src/agent/loop_run/reply_retry.rs
@@ -43,10 +43,11 @@ use super::actor_loop_flow::build_feedback_for_deterministic_content_fallback;
 use super::interrupt::InterruptFlag;
 use super::lifecycle;
 use super::model_request::{build_assistant_request_plan, request_non_streaming_assistant_reply};
+use super::reply_retry_types::{AssistantReplyRetryDecision, AssistantReplyRetryState};
 use super::spinner::{Spinner, SpinnerStopSignal};
 use super::tool_display::progress_path_display;
 use super::tool_history::successful_non_plan_repo_edit_count;
-use super::turn::{AssistantReplyRetryDecision, AssistantReplyRetryState, USER_INTERRUPT_ERROR};
+use super::turn::USER_INTERRUPT_ERROR;
 use crate::agent::prompting;
 use crate::agent::recovery;
 use crate::model_capabilities::model_capabilities;
