+++ src/agent/loop_run/turn.rs
@@ -14,37 +14,6 @@ pub(super) const TASK_CONTRACT_VERIFIER_ATTEMPT_LIMIT: usize = 3;
 pub(super) const TASK_CONTRACT_VERIFIER_REPAIR_ATTEMPT_LIMIT: usize = 6;
 pub(super) const USER_INTERRUPT_ERROR: &str = "__anvil_user_interrupt__";
 
-#[derive(Debug, Clone, Copy)]
-pub(super) struct AssistantReplyRetryState {
-    pub(super) downgraded_native_tools: bool,
-    pub(super) retries_remaining: usize,
-    pub(super) tool_call_format_retries_remaining: usize,
-    pub(super) extra_transport_retries: usize,
-    pub(super) transport_retry_count: usize,
-    pub(super) focused_edit_timeout_retry_count: usize,
-    pub(super) tool_call_format_retry_count: usize,
-}
-
-pub(super) enum AssistantReplyRetryDecision {
-    Retry,
-    ReturnReply(AssistantReply),
-    Fail(String),
-}
-
-impl AssistantReplyRetryState {
-    pub(super) fn new(chat_retries: usize, message_count: usize) -> Self {
-        Self {
-            downgraded_native_tools: false,
-            retries_remaining: chat_retries,
-            tool_call_format_retries_remaining: 2,
-            extra_transport_retries: if message_count >= 12 { 4 } else { 2 },
-            transport_retry_count: 0,
-            focused_edit_timeout_retry_count: 0,
-            tool_call_format_retry_count: 0,
-        }
-    }
-}
-
 impl Agent {
     /// Issue #646: build the active `TaskWorkspaceScope` for the current
     /// task. Pure projection of `work_root` + the active user request; no
